+++ module-menu/build.gradle
@@ -1,0 +1,9 @@
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    implementation(
+            project(':module-product'),
+            project(':module-common')
+    )
+}
