+++ module-menu/src/main/java/kitchenpos/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.domain;
 
 import static javax.persistence.GenerationType.IDENTITY;
